backend/requirements.txt
google-cloud==0.34.0
google-cloud-speech==2.7.0
moviepy==2.10
flask==2.1.3
openai==0.28
Pillow==
requests==
